+++ CHANGELOG.md
@@ -12,6 +12,8 @@ _Changes since 1.2.0_
 - Follow new appends to the end of the opened file (Shift-F)
 - Support UTF-8 with BOM and UTF-16 encodings in additional to current encoding support of UTF-8 without BOM
 - Status bar, which shows current byte range, last modified date time and current encoding
+- Shift-Click to extend selection
+- Dragging out of view to scroll and extend selection
 
 ### Changed
 - New appends to the opened file are not automatically followed anymore